+++ frontend/src/components/InstanceForm/index.ts
@@ -1,3 +1,7 @@
+import Buttons from "./Buttons.vue";
+import Form from "./Form.vue";
 import InstanceForm from "./InstanceForm.vue";
 
 export default InstanceForm;
+
+export { InstanceForm, Form, Buttons };
